5785 - Do not request 'values' when querying Dimension Options
Update dp-api-client-go to v2.159.1

When we send a query to dp-cantabular-api-ext to retrieve a list of
categories for a variable, we are also requesting the values
(i.e. the observation numbers) which is not needed here.

It was triggering the SDC rules applied on the population-type.

Resolves: 5782
Related: https://github.com/ONSdigital/dp-api-clients-go/pull/299